requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
sentence-transformers
chromadb
aiosqlite
pydantic
python-dotenv
groq
requests
pyjwt
passlib
bcrypt
pydantic[email]